Create some routes to test the caching
