+++ app/business/public-env.server.tsx
@@ -1,6 +1,6 @@
 import { publicEnvSchema } from '~/environment'
 
 const loadPublicEnv = (args: Record<string, unknown>) =>
-  publicEnvSchema.onUndeclaredKey('delete').assert(args)
+  publicEnvSchema.onUndeclaredKey('delete').assert({ ...args })
 
 export { loadPublicEnv }
