+++ app/routes/home._index.tsx
@@ -1,0 +1,5 @@
+import { Link } from '@remix-run/react'
+
+export default function Index() {
+  return <Link to="nested">Show Stripe key</Link>
+}
