+++ app/routes/home.nested.tsx
@@ -1,0 +1,5 @@
+import { getUniversalEnv } from '~/ui/public-env'
+
+export default function Index() {
+  return <h1>Stripe key: {getUniversalEnv('stripePublicKey')}</h1>
+}
